Add copyright notice and author information to source files
This commit adds a copyright notice and author information to various source files in the project. The notice includes the copyright holder, conditions for redistribution, and a disclaimer of warranties. The author information specifies Adolfo Gómez as the contributor. The changes were made to ensure compliance with licensing requirements and to provide proper attribution for the code.